Move new integration test into place and rename old integration test for reference.  New integration test uses actual pipeline for running and works on edison compute nodes with matplotlib.
